requirements.txt
Django==5.1a1
django-environ==0.11.2  # https://github.com/joke2k/django-environ
djangorestframework==3.15.1
psycopg2-binary==2.9.9
